Remove grrui and update packages.
