+++ RepoM.sln
@@ -9,8 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RepoM.App", "src\RepoM.App\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Specs", "tests\Specs\Specs.csproj", "{C91C3255-637F-45D4-93C0-6A26B4E99FC1}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GrrUi.App", "src\GrrUi.App\GrrUi.App.csproj", "{350BAB52-335E-4D14-B46A-9A6CF0786519}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RepoM.Ipc", "src\RepoM.Ipc\RepoM.Ipc.csproj", "{B38E49E8-4691-4D3B-982E-41F8E8B527E3}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Grr.App", "src\Grr.App\Grr.App.csproj", "{D5772395-C27C-4BB5-86DA-42DF7388A19B}"
@@ -117,26 +115,6 @@ Global
 		{C91C3255-637F-45D4-93C0-6A26B4E99FC1}.Release|x64.Build.0 = Release|Any CPU
 		{C91C3255-637F-45D4-93C0-6A26B4E99FC1}.Release|x86.ActiveCfg = Release|Any CPU
 		{C91C3255-637F-45D4-93C0-6A26B4E99FC1}.Release|x86.Build.0 = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|ARM.Build.0 = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|x64.Build.0 = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Debug|x86.Build.0 = Debug|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|Any CPU.Build.0 = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|ARM.ActiveCfg = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|ARM.Build.0 = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|ARM64.Build.0 = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|x64.ActiveCfg = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|x64.Build.0 = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|x86.ActiveCfg = Release|Any CPU
-		{350BAB52-335E-4D14-B46A-9A6CF0786519}.Release|x86.Build.0 = Release|Any CPU
 		{B38E49E8-4691-4D3B-982E-41F8E8B527E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B38E49E8-4691-4D3B-982E-41F8E8B527E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B38E49E8-4691-4D3B-982E-41F8E8B527E3}.Debug|ARM.ActiveCfg = Debug|Any CPU
