--- src/GrrUi.App/GrrUi.App.csproj
@@ -1,31 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <AssemblyName>grrui</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\Grr.App\ConsoleExtensions.cs" Link="ConsoleExtensions.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.2" />
-    <PackageReference Include="Terminal.Gui" Version="1.7.2" />
-    <PackageReference Include="TextCopy" Version="6.1.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\RepoM.Api\RepoM.Api.csproj" />
-    <ProjectReference Include="..\RepoM.Ipc\RepoM.Ipc.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="SendKeys">
-      <HintPath>..\..\_ref\SendKeys.exe</HintPath>
-      <Private>true</Private>
-    </Reference>
-  </ItemGroup>
-
-</Project>
