--- src/GrrUi.App/Model/RepositoryView.cs
@@ -1,33 +1,0 @@
-namespace GrrUi.App.Model;
-
-using System;
-using RepoM.Api.Git;
-using Repository = RepoM.Ipc.Repository;
-
-public class RepositoryView : IRepositoryView
-{
-    public RepositoryView(Repository repository, string displayText)
-    {
-        Repository = repository ?? throw new ArgumentNullException(nameof(repository));
-        DisplayText = displayText;
-    }
-
-    public override string ToString()
-    {
-        return DisplayText;
-    }
-
-    public Repository Repository { get; }
-
-    public string DisplayText { get; }
-
-    public string Name => Repository.Name ?? "";
-
-    public string CurrentBranch => Repository.BranchWithStatus ?? string.Empty;
-
-    public bool IsPinned { get; } = false;
-
-    public string Path => Repository.Path ?? string.Empty;
-
-    public bool HasUnpushedChanges => Repository.HasUnpushedChanges;
-}
