+++ src/RepoM.App/Controls/AcrylicContextMenu.cs
@@ -2,6 +2,7 @@ namespace RepoM.App.Controls;
 
 using System.Windows;
 using System.Windows.Controls;
+using RepoM.App.Services;
 
 public class AcrylicContextMenu : ContextMenu
 {
