+++ src/RepoM.App/Controls/AcrylicMenuItem.cs
@@ -4,6 +4,7 @@ using System;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
+using RepoM.App.Services;
 
 public class AcrylicMenuItem : MenuItem
 {
