+++ src/RepoM.App/NotifyIconViewModel.cs
@@ -1,8 +1,8 @@
 namespace RepoM.App;
 
-using System;
 using System.Windows;
 using System.Windows.Input;
+using RepoM.App.Services;
 
 /// <summary>
 /// Provides bindable properties and commands for the NotifyIcon. In this sample, the
@@ -43,30 +43,4 @@ public class NotifyIconViewModel
             {
                 CommandAction = () => Application.Current.Shutdown(),
             };
-}
-
-/// <summary>
-/// Simplistic delegate command for the demo.
-/// </summary>
-public class DelegateCommand : ICommand
-{
-    public Action? CommandAction { get; set; }
-
-    public Func<bool>? CanExecuteFunc { get; set; }
-
-    public void Execute(object? parameter)
-    {
-        CommandAction?.Invoke();
-    }
-
-    public bool CanExecute(object? parameter)
-    {
-        return CanExecuteFunc?.Invoke() ?? true;
-    }
-
-    public event EventHandler? CanExecuteChanged
-    {
-        add => CommandManager.RequerySuggested += value;
-        remove => CommandManager.RequerySuggested -= value;
-    }
 }
