+++ src/RepoM.App/Services/AcrylicHelper.cs
@@ -1,4 +1,4 @@
-namespace RepoM.App;
+namespace RepoM.App.Services;
 
 using System;
 using System.Windows;
