+++ src/RepoM.App/Services/WindowsCompositionHelper.cs
@@ -1,4 +1,4 @@
-namespace RepoM.App;
+namespace RepoM.App.Services;
 
 using System;
 using System.Drawing.Printing;
