+++ src/RepoM.Plugin.IpcService/RepoM.Plugin.IpcService.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.2" />
-    <PackageReference Include="NetMQ" Version="4.0.1.9" />
+    <PackageReference Include="NetMQ" Version="4.0.1.10" />
     <PackageReference Include="SimpleInjector" Version="5.4.1" />
     <PackageReference Include="SimpleInjector.Packaging" Version="5.0.0" />
   </ItemGroup>
