+++ tests/Specs/Specs.csproj
@@ -18,7 +18,7 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
     <PackageReference Include="Microsoft.TestPlatform.TestHost" Version="17.3.2" />
     <PackageReference Include="Moq" Version="4.18.2" />
-    <PackageReference Include="NetMQ" Version="4.0.1.9" />
+    <PackageReference Include="NetMQ" Version="4.0.1.10" />
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="6.0.0" />
